+++ test/test_permutation_importance.py
@@ -4,10 +4,10 @@ from sklearn.linear_model import LinearRegression, LogisticRegression
 from sklearn.metrics import log_loss
 from sklearn.model_selection import train_test_split
 
-from hidimstat.permutation_importance import PermutationImportance
+from hidimstat.permutation_importance_function import permutation_importance
 
 
-def test_permutation_importance(linear_scenario):
+def test_permutation_importance_no_fitting(linear_scenario):
     X, y, beta = linear_scenario
     important_features = np.where(beta != 0)[0]
     non_important_features = np.where(beta == 0)[0]
@@ -15,9 +15,9 @@ def test_permutation_importance(linear_scenario):
     X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=0)
 
     regression_model = LinearRegression()
-    regression_model.fit(X_train, y_train)
-
-    pi = PermutationImportance(
+    importance, list_loss_j, loss_reference = permutation_importance(
+        X_test,
+        y_test,
         estimator=regression_model,
         n_permutations=20,
         method="predict",
@@ -25,14 +25,6 @@ def test_permutation_importance(linear_scenario):
         n_jobs=1,
     )
 
-    pi.fit(
-        X_train,
-        y_train,
-        groups=None,
-    )
-    vim = pi.score(X_test, y_test)
-
-    importance = vim["importance"]
     assert importance.shape == (X.shape[1],)
     assert (
         importance[important_features].mean()
@@ -46,45 +38,83 @@ def test_permutation_importance(linear_scenario):
     }
     X_df = pd.DataFrame(X, columns=[f"col_{i}" for i in range(X.shape[1])])
     X_train_df, X_test_df, y_train, y_test = train_test_split(X_df, y, random_state=0)
-    regression_model.fit(X_train_df, y_train)
-    pi = PermutationImportance(
+    regression_model = LinearRegression()
+    importance, list_loss_j, loss_reference = permutation_importance(
+        X_test_df,
+        y_test,
         estimator=regression_model,
         n_permutations=20,
         method="predict",
         random_state=0,
         n_jobs=1,
-    )
-    pi.fit(
-        X_train_df,
-        y_train,
         groups=groups,
     )
-    vim = pi.score(X_test_df, y_test)
 
-    importance = vim["importance"]
+    assert importance[0].mean() > importance[1].mean()
+
+    # Same with groups
+    groups = {
+        "group_0": [i for i in important_features],
+        "the_group_1": [i for i in non_important_features],
+    }
+    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=0)
+    regression_model = LinearRegression()
+    importance, list_loss_j, loss_reference = permutation_importance(
+        X_test,
+        y_test,
+        estimator=regression_model,
+        n_permutations=20,
+        method="predict",
+        random_state=0,
+        n_jobs=1,
+        groups=groups,
+    )
+
     assert importance[0].mean() > importance[1].mean()
 
     # Classification case
-    y_clf = np.where(y > np.median(y), 1, 0)
+    y_clf = np.zeros_like(y)
+    for i, quantile in enumerate(np.arange(0.2, 0.8, 0.2)):
+        y_clf[np.where(y > np.quantile(y, quantile))] = i
     _, _, y_train_clf, y_test_clf = train_test_split(X, y_clf, random_state=0)
     logistic_model = LogisticRegression()
-    logistic_model.fit(X_train, y_train_clf)
 
-    pi_clf = PermutationImportance(
+    importance_clf, list_loss_j, loss_reference = permutation_importance(
+        X_test,
+        y_test_clf,
         estimator=logistic_model,
         n_permutations=20,
         method="predict_proba",
         random_state=0,
         n_jobs=1,
         loss=log_loss,
-    )
-
-    pi_clf.fit(
-        X_train,
-        y_train_clf,
         groups=None,
     )
-    vim_clf = pi_clf.score(X_test, y_test_clf)
 
-    importance_clf = vim_clf["importance"]
     assert importance_clf.shape == (X.shape[1],)
+
+
+def test_with_fitting(linear_scenario):
+    X, y, beta = linear_scenario
+    important_features = np.where(beta != 0)[0]
+    non_important_features = np.where(beta == 0)[0]
+
+    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=0)
+
+    regression_model = LinearRegression()
+    regression_model.fit(X_train, y_train)
+    importance, list_loss_j, loss_reference = permutation_importance(
+        X_test,
+        y_test,
+        estimator=regression_model,
+        n_permutations=20,
+        method="predict",
+        random_state=0,
+        n_jobs=1,
+    )
+
+    assert importance.shape == (X.shape[1],)
+    assert (
+        importance[important_features].mean()
+        > importance[non_important_features].mean()
+    )
